Add Schoolable concern
This adds a concern which handles common logic related to models that
belong to a school and, if not, have the ability to capture whether the
patient is home-schooled.
